refactor: Componente BarraChart
